RN-501: Moved OptionSetModel to @tupaia/database
