+++ packages/central-server/src/database/models/OptionSet.js
@@ -3,33 +3,9 @@
  * Copyright (c) 2017 Beyond Essential Systems Pty Ltd
  */
 
-import { DatabaseModel, DatabaseType, TYPES } from '@tupaia/database';
+import { OptionSetModel as CommonOptionSetModel } from '@tupaia/database';
 
-class OptionSetType extends DatabaseType {
-  static databaseType = TYPES.OPTION_SET;
-
-  async getSurveyIds() {
-    const surveyScreens = await this.database.executeSql(
-      `
-      SELECT survey_screen.survey_id
-      FROM survey_screen
-      INNER JOIN survey_screen_component
-        ON survey_screen_component.screen_id = survey_screen.id
-      INNER JOIN question
-        ON question.id = survey_screen_component.question_id
-      WHERE question.option_set_id = ?
-    `,
-      this.id,
-    );
-    return surveyScreens.map(s => s.survey_id);
-  }
-}
-
-export class OptionSetModel extends DatabaseModel {
-  get DatabaseTypeClass() {
-    return OptionSetType;
-  }
-
+export class OptionSetModel extends CommonOptionSetModel {
   meditrakConfig = {
     minAppVersion: '1.7.92',
   };
