+++ packages/database/src/modelClasses/OptionSet.js
@@ -1,0 +1,34 @@
+/**
+ * Tupaia MediTrak
+ * Copyright (c) 2017 Beyond Essential Systems Pty Ltd
+ */
+
+import { DatabaseModel } from '../DatabaseModel';
+import { DatabaseType } from '../DatabaseType';
+import { TYPES } from '../types';
+
+class OptionSetType extends DatabaseType {
+  static databaseType = TYPES.OPTION_SET;
+
+  async getSurveyIds() {
+    const surveyScreens = await this.database.executeSql(
+      `
+       SELECT survey_screen.survey_id
+       FROM survey_screen
+       INNER JOIN survey_screen_component
+         ON survey_screen_component.screen_id = survey_screen.id
+       INNER JOIN question
+         ON question.id = survey_screen_component.question_id
+       WHERE question.option_set_id = ?
+     `,
+      this.id,
+    );
+    return surveyScreens.map(s => s.survey_id);
+  }
+}
+
+export class OptionSetModel extends DatabaseModel {
+  get DatabaseTypeClass() {
+    return OptionSetType;
+  }
+}
